+++ lib/sbmt/outbox/engine.rb
@@ -8,7 +8,8 @@ module Sbmt
       isolate_namespace Sbmt::Outbox
 
       config.outbox = ActiveSupport::OrderedOptions.new.tap do |c|
-        c.base_class = "ApplicationRecord"
+        c.active_record_base_class = "ApplicationRecord"
+        c.active_job_base_class = "ApplicationJob"
         c.error_tracker = "Sbmt::Outbox::ErrorTracker"
         c.outbox_item_classes = []
         c.inbox_item_classes = []