+++ spec/jobs/sbmt/outbox/base_delete_stale_items_job_spec.rb
@@ -16,24 +16,21 @@ describe Sbmt::Outbox::BaseDeleteStaleItemsJob do
 
   describe ".enqueue" do
     it "enqueue all item classes" do
-      expect { job_class.enqueue }.to change { job_class.jobs.size }.by(1)
-      expect(job_class.jobs.last["args"]).to contain_exactly("OutboxItem")
+      expect { job_class.enqueue }.to have_enqueued_job(job_class).with("OutboxItem")
     end
   end
 
   it "deletes item" do
-    job_class.perform_async("OutboxItem")
-
-    expect { job_class.drain }.to change(OutboxItem, :count).by(-1)
+    expect { job_class.perform_now("OutboxItem") }
+      .to change(OutboxItem, :count).by(-1)
   end
 
   context "when item is too yang" do
     let(:created_at) { 1.hour.ago }
 
     it "doesn't delete item" do
-      job_class.perform_async("OutboxItem")
-
-      expect { job_class.drain }.not_to change(OutboxItem, :count)
+      expect { job_class.perform_now("OutboxItem") }
+        .not_to change(OutboxItem, :count)
     end
   end
 end