+++ lib/sbmt/outbox/engine.rb
@@ -37,9 +37,8 @@ module Sbmt
       end
 
       rake_tasks do
-        load "sbmt/outbox/tasks/process_outbox_items.rake"
-        load "sbmt/outbox/tasks/retry_failed_outbox_items.rake"
-        load "sbmt/outbox/tasks/delete_failed_outbox_items.rake"
+        load "sbmt/outbox/tasks/retry_failed_items.rake"
+        load "sbmt/outbox/tasks/delete_failed_items.rake"
       end
     end
   end