+++ spec/fabricators/outbox_item_fabricator.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-Fabricator(:outbox_item, from: "OutboxItem") do
-  payload { "test" }
-  event_key { sequence(:event_key) }
-  bucket { 0 }
-end