+++ spec/factories/inbox_item_factory.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+FactoryBot.define do
+  factory :inbox_item, class: "InboxItem" do
+    payload { "test" }
+    sequence(:event_key)
+    bucket { 0 }
+  end
+end