+++ spec/interactors/sbmt/outbox/process_item_spec.rb
@@ -30,7 +30,7 @@ describe Sbmt::Outbox::ProcessItem do
 
     context "when outbox item is not in pending state" do
       let(:outbox_item) do
-        Fabricate(
+        create(
           :outbox_item,
           status: Sbmt::Outbox::BaseItem.statuses[:failed]
         )
@@ -45,7 +45,7 @@ describe Sbmt::Outbox::ProcessItem do
     end
 
     context "when there is no any transport" do
-      let!(:outbox_item) { Fabricate(:outbox_item) }
+      let!(:outbox_item) { create(:outbox_item) }
 
       before do
         allow_any_instance_of(OutboxItem).to receive(:transports).and_return(nil)
@@ -73,7 +73,7 @@ describe Sbmt::Outbox::ProcessItem do
     end
 
     context "when outbox item produce to transport successfully" do
-      let!(:outbox_item) { Fabricate(:outbox_item) }
+      let!(:outbox_item) { create(:outbox_item) }
 
       it "returns success" do
         expect(Sbmt::Outbox.error_tracker).not_to receive(:error)
@@ -93,7 +93,7 @@ describe Sbmt::Outbox::ProcessItem do
     end
 
     context "when outbox item produce to transport unsuccessfully" do
-      let!(:outbox_item) { Fabricate(:outbox_item) }
+      let!(:outbox_item) { create(:outbox_item) }
 
       before do
         allow(producer).to receive(:publish).and_return(false)
@@ -139,7 +139,7 @@ describe Sbmt::Outbox::ProcessItem do
       end
 
       context "when retry process" do
-        let!(:outbox_item) { Fabricate(:outbox_item, processed_at: Time.current) }
+        let!(:outbox_item) { create(:outbox_item, processed_at: Time.current) }
 
         it "do not track process_latency" do
           expect { result }.not_to measure_yabeda_histogram(Yabeda.outbox.process_latency)
@@ -148,7 +148,7 @@ describe Sbmt::Outbox::ProcessItem do
     end
 
     context "when item processing raised exception" do
-      let!(:outbox_item) { Fabricate(:outbox_item) }
+      let!(:outbox_item) { create(:outbox_item) }
 
       before do
         allow(producer).to receive(:publish).and_raise("boom")
@@ -180,7 +180,7 @@ describe Sbmt::Outbox::ProcessItem do
     end
 
     context "when item processing returning failure" do
-      let!(:outbox_item) { Fabricate(:outbox_item) }
+      let!(:outbox_item) { create(:outbox_item) }
 
       before do
         allow(producer).to receive(:publish)
@@ -202,7 +202,7 @@ describe Sbmt::Outbox::ProcessItem do
     end
 
     context "when outbox item has many transports" do
-      let!(:outbox_item) { Fabricate(:outbox_item) }
+      let!(:outbox_item) { create(:outbox_item) }
 
       before do
         allow_any_instance_of(OutboxItem).to receive(:transports).and_return([producer, HttpOrderSender])
@@ -215,7 +215,7 @@ describe Sbmt::Outbox::ProcessItem do
     end
 
     context "when outbox item has custom payload builder" do
-      let!(:outbox_item) { Fabricate(:outbox_item) }
+      let!(:outbox_item) { create(:outbox_item) }
 
       before do
         allow_any_instance_of(OutboxItem).to receive(:payload_builder).and_return(PayloadRenderer)
@@ -230,7 +230,7 @@ describe Sbmt::Outbox::ProcessItem do
     end
 
     context "when checking retry strategies" do
-      let!(:outbox_item) { Fabricate(:outbox_item) }
+      let!(:outbox_item) { create(:outbox_item) }
       let(:max_retries) { 1 }
 
       before do
@@ -248,7 +248,7 @@ describe Sbmt::Outbox::ProcessItem do
 
       context "with the next processing time is greater than the current time" do
         let!(:outbox_item) do
-          Fabricate(:outbox_item, processed_at: 1.hour.from_now)
+          create(:outbox_item, processed_at: 1.hour.from_now)
         end
 
         it "doesn't increment errors count" do
@@ -262,7 +262,7 @@ describe Sbmt::Outbox::ProcessItem do
 
       context "with the next processing time is less than the current time" do
         let!(:outbox_item) do
-          Fabricate(:outbox_item, processed_at: 1.hour.ago)
+          create(:outbox_item, processed_at: 1.hour.ago)
         end
 
         it "increment errors count" do
@@ -276,11 +276,11 @@ describe Sbmt::Outbox::ProcessItem do
 
       context "when retry strategy discards item" do
         let!(:outbox_item) do
-          Fabricate(:outbox_item, processed_at: 1.hour.ago)
+          create(:outbox_item, processed_at: 1.hour.ago)
         end
 
         let!(:outbox_item_2) do
-          Fabricate(:outbox_item, status: :delivered, event_key: outbox_item.event_key)
+          create(:outbox_item, status: :delivered, event_key: outbox_item.event_key)
         end
 
         it "discards processing" do
@@ -291,7 +291,7 @@ describe Sbmt::Outbox::ProcessItem do
 
       context "when retry strategy returns unknown error" do
         let!(:outbox_item) do
-          Fabricate(:outbox_item, processed_at: 1.hour.ago)
+          create(:outbox_item, processed_at: 1.hour.ago)
         end
 
         before do