+++ lib/sbmt/outbox/middleware/open_telemetry/tracing_item_process_middleware.rb
@@ -5,18 +5,18 @@ module Sbmt
     module Middleware
       module OpenTelemetry
         class TracingItemProcessMiddleware
-          def call(_job, item_id, options = {})
+          def call(item)
             return yield unless defined?(::OpenTelemetry)
 
-            item_class = options[:item_class]
-            item_options = options[:options] || {}
+            item_class = item.class
+            item_options = item.options || {}
             item_options_headers = item_options[:headers] || item_options["headers"]
 
             return yield unless item_class && item_options_headers
 
             span_attributes = {
               "messaging.system" => "outbox",
-              "messaging.outbox.item_id" => item_id.to_s,
+              "messaging.outbox.item_id" => item.id.to_s,
               "messaging.outbox.box_type" => item_class.box_type.to_s,
               "messaging.outbox.box_name" => item_class.box_name,
               "messaging.outbox.owner" => item_class.owner,