+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - make ErrorTracker to send errors to Sentry (if available)
 - re-plugged Schked as an optional dependency
 - replaced Sidekiq with ActiveJob. Don't forget to set `config.active_job.queue_adapter = :sidekiq`
+- accept box names as arguments in maintenance tasks, ex: `rake outbox:retry_failed_items[some/box_item]`
 
 ### Fixed